Add extraction logic.
Two Problems
Problem 1:
- I don't actually think this is the right function to extract.
I think we can do "better" and have a harness around game_engine that is capable of handling the placement of the cards.
In particular, I think that it is BAD when you have the following changes on the game-board:

1 2 3 4
5 6 7 8
9 A B C

Suppose 2 6 A are a match. Then when the user clears them by selecting them, the game will collect those cards

1 3 4 5
7 8 9 B
C D E F

It seems bad that all these cards have shifted! We should "keep track" of card placement somehow. And while you
could obviously do so in OCaml, I see no reason not to ALSO verify these new display methods. So I think we
shall, and the method that OCaml "plugs into" will probably be from _that_ module instead.

Problem 2:
I've only written the extraction code! I haven't plugged into the harness (which doesn't exist yet) and I certainly
haven't compiled the OCaml to JS yet. That is something people know how to do, and I've done it before, but never in
a project like this. Bear with me.