Working prototype.
This is really ugly and I want to redo large portions of it; but it's a good proof of concept.

First, the cards are ugly.
Second, I'm not sure the "shrinking" of the board works properly.
Third, we need to add ways for the Msg to show to the user.
Fourth, I want to move as much as possible of the UI logic into Rocq.
Fifth, any organizational improvements to the code would be welcome.
Sixth, I'm not really sure I'm making the right choices in Extract.v; things feel haphazard.